+++ examples/brownfield/main.tf
@@ -31,6 +31,7 @@ provider "sapintegrationsuite" {
 resource "sapintegrationsuite_integration_package" "utilities" {
   id          = "UTILITIES"
   name        = "Utilities Integration"
+  short_text  = "Utilities integration content"
   description = "Integration content for the utilities line of business"
 }