+++ examples/greenfield/main.tf
@@ -64,6 +64,7 @@ resource "sapintegrationsuite_access_policy" "utilities" {
 resource "sapintegrationsuite_integration_package" "utilities" {
   id          = "UTILITIES"
   name        = "Utilities Integration"
+  short_text  = "Utilities integration content"
   description = "Integration content for the utilities line of business"
 }