+++ examples/resources/sapintegrationsuite_integration_package/resource.tf
@@ -1,5 +1,6 @@
 resource "sapintegrationsuite_integration_package" "utilities" {
   id          = "UTILITIES"
   name        = "Utilities Integration"
+  short_text  = "Utilities integration content"
   description = "Integration content for the utilities line of business"
 }